Conflicts:
	basestation/src/blockWorld.cpp
	basestation/src/robot.cpp
	include/anki/cozmo/basestation/robot.h
	include/anki/cozmo/robot/cozmoConfig.h

+++ coretech/common/robot/src/utilities.cpp
@@ -99,7 +99,7 @@ namespace Anki
         }
       }
 
-      const s32 stringLength = strlen(text);
+      const s32 stringLength = static_cast<s32>(strlen(text));
 
       cv::Point curOrg = org;
       for(s32 iChar=0; iChar<stringLength; iChar++) {
